os/memstore: move MemStore into os/memstore/*
